Refactor main.py
This commit refactors the main entrance point to LUGACH, found in
main.py, to streamline adding or removing apps. Now, to add an app, all
the developer has to do is create a new file in the apps module and
update __apps__ in __init__.py. LUGACH will take care of displaying the
name of the app and running it accordingly.