Add final checks for LH apps
This commit adds a conditional in the two apps that use Lighthouse (Post
Final Grades and Update Attendance Verification) which checks to see if
the user would actually like to run the application before proceeding.
This ensures that the user does not accidentally change sensitive
information in Lighthouse without being given ample notice.