Refactor authentication handling
This commit refactors the code that handles authentication for Canvas,
Top Hat, and Lighthouse. Lighthouse now uses the system credential
manager to handle tokens since storing passwords in an .env file was
probably the reason Adam and Eve were banished from the Garden of Eden.
Top Hat also allows the user to enter a new token if they have not
entered one or if their old one has expired.