Revert .env handling changes
This commit reverses changes from a previous commit to the way that .env
files are handled. Now the programs use a file called ".env" again, as
opposed to using one called "_.env".

This change may break functionality for Apple devices. I tremble in the
face of what I have done.